Move PrivacyPropagator to propagation Package and Add Documentation
