+++ docs/api/java/org/apache/sysds/runtime/privacy/class-use/PrivacyPropagator.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class org.apache.sysds.runtime.privacy.PrivacyPropagator (SystemDS 2.0.0-SNAPSHOT API)</title>
+<title>Uses of Class org.apache.sysds.runtime.privacy.propagation.PrivacyPropagator (SystemDS 2.0.0-SNAPSHOT API)</title>
 <link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../../script.js"></script>
 </head>
@@ -12,7 +12,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class org.apache.sysds.runtime.privacy.PrivacyPropagator (SystemDS 2.0.0-SNAPSHOT API)";
+            parent.document.title="Uses of Class org.apache.sysds.runtime.privacy.propagation.PrivacyPropagator (SystemDS 2.0.0-SNAPSHOT API)";
         }
     }
     catch(err) {
@@ -70,9 +70,9 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class org.apache.sysds.runtime.privacy.PrivacyPropagator" class="title">Uses of Class<br>org.apache.sysds.runtime.privacy.PrivacyPropagator</h2>
+<h2 title="Uses of Class org.apache.sysds.runtime.privacy.propagation.PrivacyPropagator" class="title">Uses of Class<br>org.apache.sysds.runtime.privacy.propagation.PrivacyPropagator</h2>
 </div>
-<div class="classUseContainer">No usage of org.apache.sysds.runtime.privacy.PrivacyPropagator</div>
+<div class="classUseContainer">No usage of org.apache.sysds.runtime.privacy.propagation.PrivacyPropagator</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
