+++ src/main/java/org/apache/sysds/runtime/privacy/propagation/MatrixMultiplicationPropagator.java
@@ -6,19 +6,40 @@ import org.apache.sysds.runtime.privacy.finegrained.FineGrainedPrivacy;
 
 import java.util.stream.Stream;
 
+/**
+ * Used for propagating constraints in a matrix multiplication.
+ */
 public abstract class MatrixMultiplicationPropagator implements Propagator {
 
 	MatrixBlock input1, input2;
 	PrivacyConstraint privacyConstraint1, privacyConstraint2;
 
+	/**
+	 * Constructor for empty instance.
+	 * The fields can later be set with the setFields method.
+	 */
 	public MatrixMultiplicationPropagator(){};
 
+	/**
+	 * Constructs the propagator and initializes the fields used for propagation.
+	 * @param input1 left-hand input in matrix multiplication.
+	 * @param privacyConstraint1 privacy constraint of left-hand input in matrix multiplication
+	 * @param input2 right-hand input in matrix multiplication
+	 * @param privacyConstraint2 privacy constraint of right-hand input in matrix multiplication
+	 */
 	public MatrixMultiplicationPropagator(
 		MatrixBlock input1, PrivacyConstraint privacyConstraint1,
 		MatrixBlock input2, PrivacyConstraint privacyConstraint2){
 		setFields(input1, privacyConstraint1, input2, privacyConstraint2);
 	}
 
+	/**
+	 * Sets all fields of propagator.
+	 * @param input1 left-hand input in matrix multiplication.
+	 * @param privacyConstraint1 privacy constraint of left-hand input in matrix multiplication
+	 * @param input2 right-hand input in matrix multiplication
+	 * @param privacyConstraint2 privacy constraint of right-hand input in matrix multiplication
+	 */
 	public void setFields(
 		MatrixBlock input1, PrivacyConstraint privacyConstraint1,
 		MatrixBlock input2, PrivacyConstraint privacyConstraint2){
@@ -58,6 +79,11 @@ public abstract class MatrixMultiplicationPropagator implements Propagator {
 		return mergedConstraint;
 	}
 
+	/**
+	 * Gets the operator types of all rows of the left-hand input in the matrix multiplication.
+	 * An operator type defines if the row will result in an aggregation or not.
+	 * @return array of operator types representing the rows of the left-hand input in the matrix multiplication
+	 */
 	public OperatorType[] getOperatorTypesRow(){
 		OperatorType[] operatorTypes = new OperatorType[input1.getNumRows()];
 		for (int i = 0; i < input1.getNumRows(); i++) {
@@ -67,6 +93,11 @@ public abstract class MatrixMultiplicationPropagator implements Propagator {
 		return operatorTypes;
 	}
 
+	/**
+	 * Gets the operator types of all columns of the right-hand input in the matrix multiplication.
+	 * An operator type defines if the column will result in an aggregation or not.
+	 * @return array of operator types representing the columns of the right-hand input in the matrix multiplication.
+	 */
 	public OperatorType[] getOperatorTypesCol(){
 		OperatorType[] operatorTypes = new OperatorType[input2.getNumColumns()];
 		for (int j = 0; j < input2.getNumColumns(); j++) {
@@ -89,6 +120,15 @@ public abstract class MatrixMultiplicationPropagator implements Propagator {
 			return OperatorType.Aggregate;
 	}
 
+	/**
+	 * An abstract method that generates the fine-grained privacy constraints based on the input
+	 * and puts it in the mergedFineGrainedConstraints.
+	 * @param mergedFineGrainedConstraints FineGrainedPrivacy instance in which the output privacy constraints are put
+	 * @param rowPrivacy privacy constraints of the rows of the left-hand input in the matrix multiplication
+	 * @param colPrivacy privacy constraints of the columns of the right-hand input in the matrix multiplication
+	 * @param operatorTypes1 operator types of the left-hand input in the matrix multiplication
+	 * @param operatorTypes2 operator types of the right-hand input in the matrix multiplication
+	 */
 	protected abstract void generateFineGrainedConstraints(FineGrainedPrivacy mergedFineGrainedConstraints,
 		PrivacyConstraint.PrivacyLevel[] rowPrivacy, PrivacyConstraint.PrivacyLevel[] colPrivacy,
 		OperatorType[] operatorTypes1, OperatorType[] operatorTypes2);
