+++ src/main/java/org/apache/sysds/runtime/privacy/propagation/MatrixMultiplicationPropagatorNaive.java
@@ -3,9 +3,13 @@ package org.apache.sysds.runtime.privacy.propagation;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
 import org.apache.sysds.runtime.privacy.PrivacyConstraint;
 import org.apache.sysds.runtime.privacy.PrivacyConstraint.PrivacyLevel;
-import org.apache.sysds.runtime.privacy.PrivacyPropagator;
 import org.apache.sysds.runtime.privacy.finegrained.FineGrainedPrivacy;
 
+/**
+ * MatrixMultiplicationPropagator that overrides generateFineGrainedConstraints
+ * with a naive propagation of the fine-grained constraints.
+ * The output is correct, but is likely less efficient than other implementations.
+ */
 public class MatrixMultiplicationPropagatorNaive extends MatrixMultiplicationPropagator{
 
 	public MatrixMultiplicationPropagatorNaive(){
@@ -18,6 +22,11 @@ public class MatrixMultiplicationPropagatorNaive extends MatrixMultiplicationPro
 	}
 
 	@Override
+	/**
+	 * Generates fine-grained constraints and puts them in the mergedFineGrainedConstraints instance.
+	 * This implementation loops over every cell of the output matrix and sets the constraint based on the
+	 * row/column privacy constraints and the row/column operator type.
+	 */
 	protected void generateFineGrainedConstraints(FineGrainedPrivacy mergedFineGrainedConstraints,
 		PrivacyLevel[] rowPrivacy, PrivacyConstraint.PrivacyLevel[] colPrivacy,
 		OperatorType[] operatorTypes1, OperatorType[] operatorTypes2) {
