+++ src/main/java/org/apache/sysds/runtime/privacy/propagation/OperatorType.java
@@ -1,8 +1,5 @@
 package org.apache.sysds.runtime.privacy.propagation;
 
-import org.apache.sysds.runtime.matrix.data.MatrixBlock;
-import org.apache.sysds.runtime.privacy.PrivacyPropagator;
-
 public enum OperatorType {
 	Aggregate,
 	NonAggregate;
