+++ src/main/java/org/apache/sysds/runtime/privacy/propagation/Propagator.java
@@ -2,6 +2,13 @@ package org.apache.sysds.runtime.privacy.propagation;
 
 import org.apache.sysds.runtime.privacy.PrivacyConstraint;
 
+/**
+ * Interface for all propagator instances.
+ */
 public interface Propagator {
+	/**
+	 * Activates the propagation and returns the output privacy constraint.
+	 * @return output privacy constraint.
+	 */
 	PrivacyConstraint propagate();
 }
